Fix question components and quiz state
